!HOTFIX: 바로기부 알고리즘 로직 오류 해결

+++ backend/src/main/java/com/e105/majoong/common/model/receiptHistory/ReceiptHistoryCustom.java
@@ -12,7 +12,7 @@ import java.util.Map;
 public interface ReceiptHistoryCustom {
     List<ReceiptCountDto> getReceiptCountByOneDay(LocalDateTime start, LocalDateTime end);
     
-    Map<String, Integer> sumDonationAmountByFarmUuidsBetween(List<String> farmUuids, LocalDateTime start, LocalDateTime end);
+    Map<String, Long> sumDonationAmountByFarmUuidsBetween(List<String> farmUuids, LocalDateTime start, LocalDateTime end);
 
     List<MonthlyDonationUsedDto> findMonthlyDonationUsed(String farmUuid);
